requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.20.0
tqdm>=4.65.0
jinja2>=3.0.0
pyyaml>=6.0.0
